autoflex: Organize for easier use
