+++ internal/framework/flex/autoflex.go
@@ -8,15 +8,9 @@ import (
 	"fmt"
 	"reflect"
 	"strings"
-	"time"
 
 	pluralize "github.com/gertd/go-pluralize"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
-	fwtypes "github.com/hashicorp/terraform-provider-aws/internal/framework/types"
 )
 
 type ResourcePrefixCtxKey string
@@ -29,52 +23,6 @@ const (
 // Expand  = TF -->  AWS
 // Flatten = AWS --> TF
 
-// Expand "expands" a resource's "business logic" data structure,
-// implemented using Terraform Plugin Framework data types, into
-// an AWS SDK for Go v2 API data structure.
-// The resource's data structure is walked and exported fields that
-// have a corresponding field in the API data structure (and a suitable
-// target data type) are copied.
-func Expand(ctx context.Context, tfObject, apiObject any, optFns ...AutoFlexOptionsFunc) diag.Diagnostics {
-	var diags diag.Diagnostics
-	expander := &autoExpander{}
-
-	for _, optFn := range optFns {
-		optFn(expander)
-	}
-
-	diags.Append(autoFlexConvert(ctx, tfObject, apiObject, expander)...)
-	if diags.HasError() {
-		diags.AddError("AutoFlEx", fmt.Sprintf("Expand[%T, %T]", tfObject, apiObject))
-		return diags
-	}
-
-	return diags
-}
-
-// Flatten "flattens" an AWS SDK for Go v2 API data structure into
-// a resource's "business logic" data structure, implemented using
-// Terraform Plugin Framework data types.
-// The API data structure's fields are walked and exported fields that
-// have a corresponding field in the resource's data structure (and a
-// suitable target data type) are copied.
-func Flatten(ctx context.Context, apiObject, tfObject any, optFns ...AutoFlexOptionsFunc) diag.Diagnostics {
-	var diags diag.Diagnostics
-	flattener := &autoFlattener{}
-
-	for _, optFn := range optFns {
-		optFn(flattener)
-	}
-
-	diags.Append(autoFlexConvert(ctx, apiObject, tfObject, flattener)...)
-	if diags.HasError() {
-		diags.AddError("AutoFlEx", fmt.Sprintf("Flatten[%T, %T]", apiObject, tfObject))
-		return diags
-	}
-
-	return diags
-}
-
 // autoFlexer is the interface implemented by an auto-flattener or expander.
 type autoFlexer interface {
 	convert(context.Context, reflect.Value, reflect.Value) diag.Diagnostics
@@ -82,10 +30,6 @@ type autoFlexer interface {
 
 // AutoFlexOptionsFunc is a type alias for an autoFlexer functional option.
 type AutoFlexOptionsFunc func(autoFlexer)
-
-type autoExpander struct{}
-
-type autoFlattener struct{}
 
 // autoFlexConvert converts `from` to `to` using the specified auto-flexer.
 func autoFlexConvert(ctx context.Context, from, to any, flexer autoFlexer) diag.Diagnostics {
@@ -231,1328 +175,4 @@ func fieldExistsInStruct(field string, str reflect.Value) bool {
 	}
 
 	return false
-}
-
-// convert converts a single Plugin Framework value to its AWS API equivalent.
-func (expander autoExpander) convert(ctx context.Context, valFrom, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	vFrom, ok := valFrom.Interface().(attr.Value)
-	if !ok {
-		diags.AddError("AutoFlEx", fmt.Sprintf("does not implement attr.Value: %s", valFrom.Kind()))
-		return diags
-	}
-
-	// No need to set the target value if there's no source value.
-	if vFrom.IsNull() || vFrom.IsUnknown() {
-		return diags
-	}
-
-	switch vFrom := vFrom.(type) {
-	// Primitive types.
-	case basetypes.BoolValuable:
-		diags.Append(expander.bool(ctx, vFrom, vTo)...)
-		return diags
-
-	case basetypes.Float64Valuable:
-		diags.Append(expander.float64(ctx, vFrom, vTo)...)
-		return diags
-
-	case basetypes.Int64Valuable:
-		diags.Append(expander.int64(ctx, vFrom, vTo)...)
-		return diags
-
-	case basetypes.StringValuable:
-		diags.Append(expander.string(ctx, vFrom, vTo)...)
-		return diags
-
-	// Aggregate types.
-	case basetypes.ListValuable:
-		diags.Append(expander.list(ctx, vFrom, vTo)...)
-		return diags
-
-	case basetypes.MapValuable:
-		diags.Append(expander.map_(ctx, vFrom, vTo)...)
-		return diags
-
-	case basetypes.SetValuable:
-		diags.Append(expander.set(ctx, vFrom, vTo)...)
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from": vFrom.Type(ctx),
-		"to":   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// bool copies a Plugin Framework Bool(ish) value to a compatible AWS API value.
-func (expander autoExpander) bool(ctx context.Context, vFrom basetypes.BoolValuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToBoolValue(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch vTo.Kind() {
-	case reflect.Bool:
-		//
-		// types.Bool -> bool.
-		//
-		vTo.SetBool(v.ValueBool())
-		return diags
-
-	case reflect.Ptr:
-		switch vTo.Type().Elem().Kind() {
-		case reflect.Bool:
-			//
-			// types.Bool -> *bool.
-			//
-			vTo.Set(reflect.ValueOf(v.ValueBoolPointer()))
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from": vFrom.Type(ctx),
-		"to":   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// float64 copies a Plugin Framework Float64(ish) value to a compatible AWS API value.
-func (expander autoExpander) float64(ctx context.Context, vFrom basetypes.Float64Valuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToFloat64Value(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch vTo.Kind() {
-	case reflect.Float32, reflect.Float64:
-		//
-		// types.Float32/types.Float64 -> float32/float64.
-		//
-		vTo.SetFloat(v.ValueFloat64())
-		return diags
-
-	case reflect.Ptr:
-		switch vTo.Type().Elem().Kind() {
-		case reflect.Float32:
-			//
-			// types.Float32/types.Float64 -> *float32.
-			//
-			to := float32(v.ValueFloat64())
-			vTo.Set(reflect.ValueOf(&to))
-			return diags
-
-		case reflect.Float64:
-			//
-			// types.Float32/types.Float64 -> *float64.
-			//
-			vTo.Set(reflect.ValueOf(v.ValueFloat64Pointer()))
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from": vFrom.Type(ctx),
-		"to":   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// int64 copies a Plugin Framework Int64(ish) value to a compatible AWS API value.
-func (expander autoExpander) int64(ctx context.Context, vFrom basetypes.Int64Valuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToInt64Value(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch vTo.Kind() {
-	case reflect.Int32, reflect.Int64:
-		//
-		// types.Int32/types.Int64 -> int32/int64.
-		//
-		vTo.SetInt(v.ValueInt64())
-		return diags
-
-	case reflect.Ptr:
-		switch vTo.Type().Elem().Kind() {
-		case reflect.Int32:
-			//
-			// types.Int32/types.Int64 -> *int32.
-			//
-			to := int32(v.ValueInt64())
-			vTo.Set(reflect.ValueOf(&to))
-			return diags
-
-		case reflect.Int64:
-			//
-			// types.Int32/types.Int64 -> *int64.
-			//
-			vTo.Set(reflect.ValueOf(v.ValueInt64Pointer()))
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from": vFrom.Type(ctx),
-		"to":   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// string copies a Plugin Framework String(ish) value to a compatible AWS API value.
-func (expander autoExpander) string(ctx context.Context, vFrom basetypes.StringValuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToStringValue(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch vTo.Kind() {
-	case reflect.String:
-		//
-		// types.String -> string.
-		//
-		vTo.SetString(v.ValueString())
-		return diags
-	case reflect.Struct:
-		//
-		// fwtypes.Timestamp --> time.Time
-		//
-		if t, ok := vFrom.(fwtypes.Timestamp); ok {
-			vTo.Set(reflect.ValueOf(t.ValueTimestamp()))
-			return diags
-		}
-	case reflect.Ptr:
-		switch vTo.Type().Elem().Kind() {
-		case reflect.String:
-			//
-			// types.String -> *string.
-			//
-			vTo.Set(reflect.ValueOf(v.ValueStringPointer()))
-			return diags
-		case reflect.Struct:
-			//
-			// fwtypes.Timestamp --> *time.Time
-			//
-			if t, ok := vFrom.(fwtypes.Timestamp); ok {
-				vTo.Set(reflect.ValueOf(t.ValueTimestampPointer()))
-				return diags
-			}
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from": vFrom.Type(ctx),
-		"to":   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// list copies a Plugin Framework List(ish) value to a compatible AWS API value.
-func (expander autoExpander) list(ctx context.Context, vFrom basetypes.ListValuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToListValue(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch v.ElementType(ctx).(type) {
-	case basetypes.StringTypable:
-		diags.Append(expander.listOfString(ctx, v, vTo)...)
-		return diags
-
-	case basetypes.ObjectTypable:
-		if vFrom, ok := vFrom.(fwtypes.NestedObjectValue); ok {
-			diags.Append(expander.nestedObject(ctx, vFrom, vTo)...)
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from list[%s]": v.ElementType(ctx),
-		"to":            vTo.Kind(),
-	})
-
-	return diags
-}
-
-// listOfString copies a Plugin Framework ListOfString(ish) value to a compatible AWS API value.
-func (expander autoExpander) listOfString(ctx context.Context, vFrom basetypes.ListValue, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch vTo.Kind() {
-	case reflect.Slice:
-		switch tSliceElem := vTo.Type().Elem(); tSliceElem.Kind() {
-		case reflect.String:
-			//
-			// types.List(OfString) -> []string.
-			//
-			var to []string
-			diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-			if diags.HasError() {
-				return diags
-			}
-
-			vTo.Set(reflect.ValueOf(to))
-			return diags
-
-		case reflect.Ptr:
-			switch tSliceElem.Elem().Kind() {
-			case reflect.String:
-				//
-				// types.List(OfString) -> []*string.
-				//
-				var to []*string
-				diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-			}
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from list[%s]": vFrom.ElementType(ctx),
-		"to":            vTo.Kind(),
-	})
-
-	return diags
-}
-
-// map_ copies a Plugin Framework Map(ish) value to a compatible AWS API value.
-func (expander autoExpander) map_(ctx context.Context, vFrom basetypes.MapValuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToMapValue(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch v.ElementType(ctx).(type) {
-	case basetypes.StringTypable:
-		diags.Append(expander.mapOfString(ctx, v, vTo)...)
-		return diags
-	case basetypes.ObjectTypable:
-		if vFrom, ok := vFrom.(fwtypes.ObjectMapValue); ok {
-			diags.Append(expander.objectMap(ctx, vFrom, vTo)...)
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from map[string, %s]": v.ElementType(ctx),
-		"to":                   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// mapOfString copies a Plugin Framework MapOfString(ish) value to a compatible AWS API value.
-func (expander autoExpander) mapOfString(ctx context.Context, vFrom basetypes.MapValue, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch vTo.Kind() {
-	case reflect.Map:
-		switch tMapKey := vTo.Type().Key(); tMapKey.Kind() {
-		case reflect.String: // key
-			switch tMapElem := vTo.Type().Elem(); tMapElem.Kind() {
-			case reflect.String:
-				//
-				// types.Map(OfString) -> map[string]string.
-				//
-				var to map[string]string
-				diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-
-			case reflect.Ptr:
-				switch k := tMapElem.Elem().Kind(); k {
-				case reflect.String:
-					//
-					// types.Map(OfString) -> map[string]*string.
-					//
-					var to map[string]*string
-					diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-					if diags.HasError() {
-						return diags
-					}
-
-					vTo.Set(reflect.ValueOf(to))
-					return diags
-				}
-			}
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from map[string, %s]": vFrom.ElementType(ctx),
-		"to":                   vTo.Kind(),
-	})
-
-	return diags
-}
-
-// set copies a Plugin Framework Set(ish) value to a compatible AWS API value.
-func (expander autoExpander) set(ctx context.Context, vFrom basetypes.SetValuable, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	v, d := vFrom.ToSetValue(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	switch v.ElementType(ctx).(type) {
-	case basetypes.StringTypable:
-		diags.Append(expander.setOfString(ctx, v, vTo)...)
-		return diags
-
-	case basetypes.ObjectTypable:
-		if vFrom, ok := vFrom.(fwtypes.NestedObjectValue); ok {
-			diags.Append(expander.nestedObject(ctx, vFrom, vTo)...)
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from set[%s]": v.ElementType(ctx),
-		"to":           vTo.Kind(),
-	})
-
-	return diags
-}
-
-// setOfString copies a Plugin Framework SetOfString(ish) value to a compatible AWS API value.
-func (expander autoExpander) setOfString(ctx context.Context, vFrom basetypes.SetValue, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch vTo.Kind() {
-	case reflect.Slice:
-		switch tSliceElem := vTo.Type().Elem(); tSliceElem.Kind() {
-		case reflect.String:
-			//
-			// types.Set(OfString) -> []string.
-			//
-			var to []string
-			diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-			if diags.HasError() {
-				return diags
-			}
-
-			vTo.Set(reflect.ValueOf(to))
-			return diags
-
-		case reflect.Ptr:
-			switch tSliceElem.Elem().Kind() {
-			case reflect.String:
-				//
-				// types.Set(OfString) -> []*string.
-				//
-				var to []*string
-				diags.Append(vFrom.ElementsAs(ctx, &to, false)...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-			}
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Expand; incompatible types", map[string]interface{}{
-		"from set[%s]": vFrom.ElementType(ctx),
-		"to":           vTo.Kind(),
-	})
-
-	return diags
-}
-
-// nestedObject copies a Plugin Framework NestedObjectValue value to a compatible AWS API value.
-func (expander autoExpander) nestedObject(ctx context.Context, vFrom fwtypes.NestedObjectValue, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := vTo.Type(); vTo.Kind() {
-	case reflect.Struct:
-		diags.Append(expander.nestedObjectToStruct(ctx, vFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Ptr:
-		switch tElem := tTo.Elem(); tElem.Kind() {
-		case reflect.Struct:
-			//
-			// types.List(OfObject) -> *struct.
-			//
-			diags.Append(expander.nestedObjectToStruct(ctx, vFrom, tElem, vTo)...)
-			return diags
-		}
-
-	case reflect.Slice:
-		switch tElem := tTo.Elem(); tElem.Kind() {
-		case reflect.Struct:
-			//
-			// types.List(OfObject) -> []struct.
-			//
-			diags.Append(expander.nestedObjectToSlice(ctx, vFrom, tTo, tElem, vTo)...)
-			return diags
-
-		case reflect.Ptr:
-			switch tElem := tElem.Elem(); tElem.Kind() {
-			case reflect.Struct:
-				//
-				// types.List(OfObject) -> []*struct.
-				//
-				diags.Append(expander.nestedObjectToSlice(ctx, vFrom, tTo, tElem, vTo)...)
-				return diags
-			}
-		}
-	}
-
-	diags.AddError("Incompatible types", fmt.Sprintf("nestedObject[%s] cannot be expanded to %s", vFrom.Type(ctx).(attr.TypeWithElementType).ElementType(), vTo.Kind()))
-	return diags
-}
-
-// nestedObjectToStruct copies a Plugin Framework NestedObjectValue to a compatible AWS API (*)struct value.
-func (expander autoExpander) nestedObjectToStruct(ctx context.Context, vFrom fwtypes.NestedObjectValue, tStruct reflect.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	// Get the nested Object as a pointer.
-	from, d := vFrom.ToObjectPtr(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	// Create a new target structure and walk its fields.
-	to := reflect.New(tStruct)
-	diags.Append(autoFlexConvertStruct(ctx, from, to.Interface(), expander)...)
-	if diags.HasError() {
-		return diags
-	}
-
-	// Set value (or pointer).
-	if vTo.Type().Kind() == reflect.Struct {
-		vTo.Set(to.Elem())
-	} else {
-		vTo.Set(to)
-	}
-
-	return diags
-}
-
-// nestedObjectToSlice copies a Plugin Framework NestedObjectValue to a compatible AWS API [](*)struct value.
-func (expander autoExpander) nestedObjectToSlice(ctx context.Context, vFrom fwtypes.NestedObjectValue, tSlice, tElem reflect.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	// Get the nested Objects as a slice.
-	from, d := vFrom.ToObjectSlice(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	// Create a new target slice and expand each element.
-	f := reflect.ValueOf(from)
-	n := f.Len()
-	t := reflect.MakeSlice(tSlice, n, n)
-	for i := 0; i < n; i++ {
-		// Create a new target structure and walk its fields.
-		target := reflect.New(tElem)
-		diags.Append(autoFlexConvertStruct(ctx, f.Index(i).Interface(), target.Interface(), expander)...)
-		if diags.HasError() {
-			return diags
-		}
-
-		// Set value (or pointer) in the target slice.
-		if vTo.Type().Elem().Kind() == reflect.Struct {
-			t.Index(i).Set(target.Elem())
-		} else {
-			t.Index(i).Set(target)
-		}
-	}
-
-	vTo.Set(t)
-
-	return diags
-}
-
-// objectMap copies a Plugin Framework ObjectMapValue value to a compatible AWS API value.
-func (expander autoExpander) objectMap(ctx context.Context, vFrom fwtypes.ObjectMapValue, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := vTo.Type(); vTo.Kind() {
-	case reflect.Map:
-		switch tElem := tTo.Elem(); tElem.Kind() {
-		case reflect.Ptr:
-			diags.Append(expander.mappedObjectToStruct(ctx, vFrom, tElem, vTo)...)
-			return diags
-		case reflect.Struct:
-			diags.Append(expander.mappedObjectToStruct(ctx, vFrom, tElem, vTo)...)
-			return diags
-		}
-	case reflect.Ptr:
-		switch tElem := tTo.Elem(); tElem.Kind() {
-		case reflect.Struct:
-			diags.Append(expander.mappedObjectToStruct(ctx, vFrom, tElem, vTo)...)
-			return diags
-		}
-	}
-
-	diags.AddError("Incompatible types", fmt.Sprintf("objectMap[%s] cannot be expanded to %s", vFrom.Type(ctx).(attr.TypeWithElementType).ElementType(), vTo.Kind()))
-	return diags
-}
-
-// mappedObjectToStruct copies a Plugin Framework ObjectMapValue to a compatible AWS API (*)struct value.
-func (expander autoExpander) mappedObjectToStruct(ctx context.Context, vFrom fwtypes.ObjectMapValue, tStruct reflect.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	// Get the mapped Objects as a map
-	from, d := vFrom.ToObjectMap(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	if tStruct.Kind() == reflect.Ptr {
-		tStruct = tStruct.Elem()
-	}
-
-	f := reflect.ValueOf(from)
-	m := reflect.MakeMap(vTo.Type())
-
-	for _, key := range f.MapKeys() {
-		// Create a new target map and populate it
-		target := reflect.New(tStruct)
-
-		fromInterface := f.MapIndex(key).Interface()
-		if f.MapIndex(key).Kind() == reflect.Ptr {
-			fromInterface = f.MapIndex(key).Elem().Interface()
-		}
-
-		diags.Append(autoFlexConvertStruct(ctx, fromInterface, target.Interface(), expander)...)
-		if diags.HasError() {
-			return diags
-		}
-
-		// Set value (or pointer) in the target slice.
-		if vTo.Type().Elem().Kind() == reflect.Struct {
-			m.SetMapIndex(key, target.Elem())
-		} else {
-			m.SetMapIndex(key, target)
-		}
-	}
-
-	vTo.Set(m)
-
-	return diags
-}
-
-// convert converts a single AWS API value to its Plugin Framework equivalent.
-func (flattener autoFlattener) convert(ctx context.Context, vFrom, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	valTo, ok := vTo.Interface().(attr.Value)
-	if !ok {
-		diags.AddError("AutoFlEx", fmt.Sprintf("does not implement attr.Value: %s", vTo.Kind()))
-		return diags
-	}
-
-	tTo := valTo.Type(ctx)
-	switch k := vFrom.Kind(); k {
-	case reflect.Bool:
-		diags.Append(flattener.bool(ctx, vFrom, false, tTo, vTo)...)
-		return diags
-
-	case reflect.Float32, reflect.Float64:
-		diags.Append(flattener.float(ctx, vFrom, false, tTo, vTo)...)
-		return diags
-
-	case reflect.Int32, reflect.Int64:
-		diags.Append(flattener.int(ctx, vFrom, false, tTo, vTo)...)
-		return diags
-
-	case reflect.String:
-		diags.Append(flattener.string(ctx, vFrom, false, tTo, vTo)...)
-		return diags
-
-	case reflect.Ptr:
-		diags.Append(flattener.ptr(ctx, vFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Slice:
-		diags.Append(flattener.slice(ctx, vFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Map:
-		diags.Append(flattener.map_(ctx, vFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Struct:
-		diags.Append(flattener.struct_(ctx, vFrom, false, tTo, vTo)...)
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// bool copies an AWS API bool value to a compatible Plugin Framework value.
-func (flattener autoFlattener) bool(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := tTo.(type) {
-	case basetypes.BoolTypable:
-		boolValue := types.BoolNull()
-		if !isNullFrom {
-			boolValue = types.BoolValue(vFrom.Bool())
-		}
-		v, d := tTo.ValueFromBool(ctx, boolValue)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		//
-		// bool -> types.Bool.
-		//
-		vTo.Set(reflect.ValueOf(v))
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// float copies an AWS API float value to a compatible Plugin Framework value.
-func (flattener autoFlattener) float(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := tTo.(type) {
-	case basetypes.Float64Typable:
-		float64Value := types.Float64Null()
-		if !isNullFrom {
-			float64Value = types.Float64Value(vFrom.Float())
-		}
-		v, d := tTo.ValueFromFloat64(ctx, float64Value)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		//
-		// float32/float64 -> types.Float64.
-		//
-		vTo.Set(reflect.ValueOf(v))
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// int copies an AWS API int value to a compatible Plugin Framework value.
-func (flattener autoFlattener) int(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := tTo.(type) {
-	case basetypes.Int64Typable:
-		int64Value := types.Int64Null()
-		if !isNullFrom {
-			int64Value = types.Int64Value(vFrom.Int())
-		}
-		v, d := tTo.ValueFromInt64(ctx, int64Value)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		//
-		// int32/int64 -> types.Int64.
-		//
-		vTo.Set(reflect.ValueOf(v))
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// string copies an AWS API string value to a compatible Plugin Framework value.
-func (flattener autoFlattener) string(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tTo := tTo.(type) {
-	case basetypes.StringTypable:
-		stringValue := types.StringNull()
-		if !isNullFrom {
-			stringValue = types.StringValue(vFrom.String())
-		}
-		v, d := tTo.ValueFromString(ctx, stringValue)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		//
-		// string -> types.String.
-		//
-		vTo.Set(reflect.ValueOf(v))
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-func (flattener autoFlattener) time(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo fwtypes.Timestamp, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if isNullFrom {
-		vTo.Set(reflect.ValueOf(fwtypes.TimestampNull()))
-		return diags
-	}
-
-	valTo, ok := vTo.Interface().(attr.Value)
-	if !ok {
-		diags.AddError("AutoFlEx", fmt.Sprintf("does not implement attr.Value: %s", vTo.Kind()))
-		return diags
-	}
-
-	_, ok = valTo.(fwtypes.Timestamp)
-	if !ok {
-		diags.AddError("AutoFlEx", fmt.Sprintf("time.Time flattens to fwtypes.Timestamp, not: %s", vTo.Kind()))
-		return diags
-	}
-
-	// time.Time --> fwtypes.Timestamp
-	if from, ok := vFrom.Interface().(time.Time); ok {
-		vTo.Set(reflect.ValueOf(fwtypes.TimestampValue(from.Format(time.RFC3339))))
-		return diags
-	}
-
-	// *time.Time --> fwtypes.Timestamp
-	if from, ok := vFrom.Elem().Interface().(time.Time); ok {
-		vTo.Set(reflect.ValueOf(fwtypes.TimestampValue(from.Format(time.RFC3339))))
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   vTo,
-	})
-
-	return diags
-}
-
-// ptr copies an AWS API pointer value to a compatible Plugin Framework value.
-func (flattener autoFlattener) ptr(ctx context.Context, vFrom reflect.Value, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch vElem, isNilFrom := vFrom.Elem(), vFrom.IsNil(); vFrom.Type().Elem().Kind() {
-	case reflect.Bool:
-		diags.Append(flattener.bool(ctx, vElem, isNilFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Float32, reflect.Float64:
-		diags.Append(flattener.float(ctx, vElem, isNilFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Int32, reflect.Int64:
-		diags.Append(flattener.int(ctx, vElem, isNilFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.String:
-		diags.Append(flattener.string(ctx, vElem, isNilFrom, tTo, vTo)...)
-		return diags
-
-	case reflect.Struct:
-		diags.Append(flattener.struct_(ctx, vElem, isNilFrom, tTo, vTo)...)
-		return diags
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// struct_ copies an AWS API struct value to a compatible Plugin Framework value.
-func (flattener autoFlattener) struct_(ctx context.Context, vFrom reflect.Value, isNilFrom bool, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if tTo, ok := tTo.(fwtypes.NestedObjectType); ok {
-		//
-		// *struct -> types.List(OfObject).
-		//
-		diags.Append(flattener.structToNestedObject(ctx, vFrom, isNilFrom, tTo, vTo)...)
-		return diags
-	}
-
-	interfaceAttrVal, ok := vTo.Interface().(attr.Value)
-	if !ok {
-		diags.AddError("AutoFlEx", fmt.Sprintf("does not implement attr.Value: %s", vTo.Kind()))
-		return diags
-	}
-
-	switch vToType := interfaceAttrVal.(type) {
-	case fwtypes.Timestamp:
-		diags.Append(flattener.time(ctx, vFrom, isNilFrom, vToType, vTo)...)
-		return diags
-	}
-
-	return diags
-}
-
-// slice copies an AWS API slice value to a compatible Plugin Framework value.
-func (flattener autoFlattener) slice(ctx context.Context, vFrom reflect.Value, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tSliceElem := vFrom.Type().Elem(); tSliceElem.Kind() {
-	case reflect.String:
-		switch tTo := tTo.(type) {
-		case basetypes.ListTypable:
-			//
-			// []string -> types.List(OfString).
-			//
-			if vFrom.IsNil() {
-				vTo.Set(reflect.ValueOf(types.ListNull(types.StringType)))
-				return diags
-			}
-
-			elements := make([]attr.Value, vFrom.Len())
-			for i := 0; i < vFrom.Len(); i++ {
-				elements[i] = types.StringValue(vFrom.Index(i).String())
-			}
-			list, d := types.ListValue(types.StringType, elements)
-			diags.Append(d...)
-			if diags.HasError() {
-				return diags
-			}
-
-			to, d := tTo.ValueFromList(ctx, list)
-			diags.Append(d...)
-			if diags.HasError() {
-				return diags
-			}
-
-			vTo.Set(reflect.ValueOf(to))
-			return diags
-
-		case basetypes.SetTypable:
-			//
-			// []string -> types.Set(OfString).
-			//
-			if vFrom.IsNil() {
-				vTo.Set(reflect.ValueOf(types.SetNull(types.StringType)))
-				return diags
-			}
-
-			elements := make([]attr.Value, vFrom.Len())
-			for i := 0; i < vFrom.Len(); i++ {
-				elements[i] = types.StringValue(vFrom.Index(i).String())
-			}
-			set, d := types.SetValue(types.StringType, elements)
-			diags.Append(d...)
-			if diags.HasError() {
-				return diags
-			}
-
-			to, d := tTo.ValueFromSet(ctx, set)
-			diags.Append(d...)
-			if diags.HasError() {
-				return diags
-			}
-
-			vTo.Set(reflect.ValueOf(to))
-			return diags
-		}
-
-	case reflect.Ptr:
-		switch tSliceElem.Elem().Kind() {
-		case reflect.String:
-			switch tTo := tTo.(type) {
-			case basetypes.ListTypable:
-				//
-				// []*string -> types.List(OfString).
-				//
-				if vFrom.IsNil() {
-					vTo.Set(reflect.ValueOf(types.ListNull(types.StringType)))
-					return diags
-				}
-
-				from := vFrom.Interface().([]*string)
-				elements := make([]attr.Value, len(from))
-				for i, v := range from {
-					elements[i] = types.StringPointerValue(v)
-				}
-				list, d := types.ListValue(types.StringType, elements)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				to, d := tTo.ValueFromList(ctx, list)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-
-			case basetypes.SetTypable:
-				//
-				// []string -> types.Set(OfString).
-				//
-				if vFrom.IsNil() {
-					vTo.Set(reflect.ValueOf(types.SetNull(types.StringType)))
-					return diags
-				}
-
-				from := vFrom.Interface().([]*string)
-				elements := make([]attr.Value, len(from))
-				for i, v := range from {
-					elements[i] = types.StringPointerValue(v)
-				}
-				set, d := types.SetValue(types.StringType, elements)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				to, d := tTo.ValueFromSet(ctx, set)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-			}
-
-		case reflect.Struct:
-			if tTo, ok := tTo.(fwtypes.NestedObjectType); ok {
-				//
-				// []*struct -> types.List(OfObject).
-				//
-				diags.Append(flattener.sliceOfStructNestedObject(ctx, vFrom, tTo, vTo)...)
-				return diags
-			}
-		}
-
-	case reflect.Struct:
-		if tTo, ok := tTo.(fwtypes.NestedObjectType); ok {
-			//
-			// []struct -> types.List(OfObject).
-			//
-			diags.Append(flattener.sliceOfStructNestedObject(ctx, vFrom, tTo, vTo)...)
-			return diags
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-// map_ copies an AWS API map value to a compatible Plugin Framework value.
-func (flattener autoFlattener) map_(ctx context.Context, vFrom reflect.Value, tTo attr.Type, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	switch tMapKey := vFrom.Type().Key(); tMapKey.Kind() {
-	case reflect.String:
-		switch tMapElem := vFrom.Type().Elem(); tMapElem.Kind() {
-		case reflect.Struct:
-			switch tTo := tTo.(type) {
-			case basetypes.MapTypable:
-				//
-				// map[string]struct -> fwtypes.ObjectMapOf[Object]
-				//
-				if tTo, ok := tTo.(fwtypes.ObjectMapType); ok {
-					diags.Append(flattener.structMapToObjectMap(ctx, vFrom, tTo, vTo)...)
-					return diags
-				}
-			}
-		case reflect.String:
-			switch tTo := tTo.(type) {
-			case basetypes.MapTypable:
-				//
-				// map[string]string -> types.Map(OfString).
-				//
-				if vFrom.IsNil() {
-					vTo.Set(reflect.ValueOf(types.MapNull(types.StringType)))
-					return diags
-				}
-
-				from := vFrom.Interface().(map[string]string)
-				elements := make(map[string]attr.Value, len(from))
-				for k, v := range from {
-					elements[k] = types.StringValue(v)
-				}
-				map_, d := types.MapValue(types.StringType, elements)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				to, d := tTo.ValueFromMap(ctx, map_)
-				diags.Append(d...)
-				if diags.HasError() {
-					return diags
-				}
-
-				vTo.Set(reflect.ValueOf(to))
-				return diags
-			}
-
-		case reflect.Ptr:
-			switch tMapElem.Elem().Kind() {
-			case reflect.Struct:
-				//
-				// map[string]*struct -> fwtypes.ObjectMapOf[Object]
-				//
-				if tTo, ok := tTo.(fwtypes.ObjectMapType); ok {
-					diags.Append(flattener.structMapToObjectMap(ctx, vFrom, tTo, vTo)...)
-					return diags
-				}
-			case reflect.String:
-				switch tTo := tTo.(type) {
-				case basetypes.MapTypable:
-					//
-					// map[string]*string -> types.Map(OfString).
-					//
-					if vFrom.IsNil() {
-						vTo.Set(reflect.ValueOf(types.MapNull(types.StringType)))
-						return diags
-					}
-
-					from := vFrom.Interface().(map[string]*string)
-					elements := make(map[string]attr.Value, len(from))
-					for k, v := range from {
-						elements[k] = types.StringPointerValue(v)
-					}
-					map_, d := types.MapValue(types.StringType, elements)
-					diags.Append(d...)
-					if diags.HasError() {
-						return diags
-					}
-
-					to, d := tTo.ValueFromMap(ctx, map_)
-					diags.Append(d...)
-					if diags.HasError() {
-						return diags
-					}
-
-					vTo.Set(reflect.ValueOf(to))
-					return diags
-				}
-			}
-		}
-	}
-
-	tflog.Info(ctx, "AutoFlex Flatten; incompatible types", map[string]interface{}{
-		"from": vFrom.Kind(),
-		"to":   tTo,
-	})
-
-	return diags
-}
-
-func (flattener autoFlattener) structMapToObjectMap(ctx context.Context, vFrom reflect.Value, tTo fwtypes.ObjectMapType, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if vFrom.IsNil() {
-		val, d := tTo.NullValue(ctx)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		vTo.Set(reflect.ValueOf(val))
-		return diags
-	}
-
-	to, d := tTo.New(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	t := reflect.ValueOf(to)
-
-	tStruct := t.Type().Elem()
-	if tStruct.Kind() == reflect.Ptr {
-		tStruct = tStruct.Elem()
-	}
-
-	for _, key := range vFrom.MapKeys() {
-		target := reflect.New(tStruct)
-
-		fromInterface := vFrom.MapIndex(key).Interface()
-		if vFrom.MapIndex(key).Kind() == reflect.Ptr {
-			fromInterface = vFrom.MapIndex(key).Elem().Interface()
-		}
-
-		diags.Append(autoFlexConvertStruct(ctx, fromInterface, target.Interface(), flattener)...)
-		if diags.HasError() {
-			return diags
-		}
-
-		if t.Type().Elem().Kind() == reflect.Struct {
-			t.SetMapIndex(key, target.Elem())
-		} else {
-			t.SetMapIndex(key, target)
-		}
-	}
-
-	val, d := tTo.ValueFromRawMap(ctx, to)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	vTo.Set(reflect.ValueOf(val))
-	return diags
-}
-
-// structToNestedObject copies an AWS API struct value to a compatible Plugin Framework NestedObjectValue value.
-func (flattener autoFlattener) structToNestedObject(ctx context.Context, vFrom reflect.Value, isNullFrom bool, tTo fwtypes.NestedObjectType, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if isNullFrom {
-		val, d := tTo.NullValue(ctx)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		vTo.Set(reflect.ValueOf(val))
-		return diags
-	}
-
-	// Create a new target structure and walk its fields.
-	to, d := tTo.NewObjectPtr(ctx)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	diags.Append(autoFlexConvertStruct(ctx, vFrom.Interface(), to, flattener)...)
-	if diags.HasError() {
-		return diags
-	}
-
-	// Set the target structure as a mapped Object.
-	val, d := tTo.ValueFromObjectPtr(ctx, to)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	vTo.Set(reflect.ValueOf(val))
-	return diags
-}
-
-// sliceOfStructNestedObject copies an AWS API []struct value to a compatible Plugin Framework NestedObjectValue value.
-func (flattener autoFlattener) sliceOfStructNestedObject(ctx context.Context, vFrom reflect.Value, tTo fwtypes.NestedObjectType, vTo reflect.Value) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if vFrom.IsNil() {
-		val, d := tTo.NullValue(ctx)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		vTo.Set(reflect.ValueOf(val))
-		return diags
-	}
-
-	// Create a new target slice and flatten each element.
-	n := vFrom.Len()
-	to, d := tTo.NewObjectSlice(ctx, n, n)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	t := reflect.ValueOf(to)
-	for i := 0; i < n; i++ {
-		target, d := tTo.NewObjectPtr(ctx)
-		diags.Append(d...)
-		if diags.HasError() {
-			return diags
-		}
-
-		diags.Append(autoFlexConvertStruct(ctx, vFrom.Index(i).Interface(), target, flattener)...)
-		if diags.HasError() {
-			return diags
-		}
-
-		t.Index(i).Set(reflect.ValueOf(target))
-	}
-
-	// Set the target structure as a nested Object.
-	val, d := tTo.ValueFromObjectSlice(ctx, to)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	vTo.Set(reflect.ValueOf(val))
-	return diags
 }
